Fix function "calculateLift" of the class PeakLiftFunctionImpl.

+++ python/boomer/seco/cpp/lift_functions.cpp
@@ -31,9 +31,9 @@ float64 PeakLiftFunctionImpl::calculateLift(uint32 numLabels) {
     float64 normalization;
 
     if (numLabels < peakLabel_) {
-        normalization = (numLabels - 1) / ((float64) (peakLabel_ - 1));
+        normalization = ((float64) numLabels - 1) / ((float64) peakLabel_ - 1);
     } else if (numLabels > peakLabel_) {
-        normalization = (numLabels - numLabels_) / ((float64) (numLabels_ - peakLabel_));
+        normalization = ((float64) numLabels - (float64) numLabels_) / ((float64) numLabels_ - (float64) peakLabel_);
     } else {
         return maxLift_;
     }
